Now emulates setting the frame for mw.getCurrentFrame calls to work inside modules.
Can also serve both direct-debugging of modules exactly as-is on your live MW production site (but without invoking their functions) and test script files. An example script file is provided that requires a module and calls its main function just like a template call from a page on a wiki site.

Note that you do not need to (and cannot) require("mw_runner") in the test script or it will cause an infinite loop of invoking itself.